+++ tests/column_view/cropping.c
@@ -13,7 +13,7 @@ static const size_t MAX_WIDTH = 40;
 static char print_buffer[40 + 1];
 
 static void column_line_print(const void *data, int column_id, const char buf[],
-		size_t offset, AlignType align);
+		int offset, AlignType align);
 static void column1_func(void *data, size_t buf_len, char buf[],
 		const format_info_t *info);
 static void column2_func(void *data, size_t buf_len, char buf[],
@@ -36,8 +36,8 @@ TEARDOWN()
 }
 
 static void
-column_line_print(const void *data, int column_id, const char buf[],
-		size_t offset, AlignType align)
+column_line_print(const void *data, int column_id, const char buf[], int offset,
+		AlignType align)
 {
 	memcpy(print_buffer + offset, buf, strlen(buf));
 }
